# perf/perf_pmu: Extend the test to support Power11