--- a/perf/perf_pmu.py.data/sysfs_PMU.cfg
+++ b/perf/perf_pmu.py.data/sysfs_PMU.cfg
@@ -1,0 +1,8 @@
+[power8]
+dir_list = mmcr0,mmcr1,mmcra
+[power8e]
+dir_list = mmcr0,mmcr1,mmcra
+[power9]
+dir_list = mmcr0,mmcr1,mmcra
+[power10]
+dir_list = mmcr0,mmcr1,mmcr3,mmcra