--- a/perf/perf_pmu.py.data/sysfs_PMU.cfg
+++ b/perf/perf_pmu.py.data/sysfs_PMU.cfg
@@ -6,3 +6,5 @@ dir_list = mmcr0,mmcr1,mmcra
 dir_list = mmcr0,mmcr1,mmcra
 [power10]
 dir_list = mmcr0,mmcr1,mmcr3,mmcra
+[power11]
+dir_list = mmcr0,mmcr1,mmcr3,mmcra