Extend the test case to support Power11 processor

Signed-off-by: Sachin Sant <sachinp@linux.ibm.com>