@sacsant 2024-08-08
Before the change:
 (1/8) perf/perf_pmu.py:PerfBasic.test_config_PMU_sysfs: STARTED
 (1/8) perf/perf_pmu.py:PerfBasic.test_config_PMU_sysfs:  ERROR: 'power11' (0.06 s)
 (2/8) perf/perf_pmu.py:PerfBasic.test_config_PPC_RTAS: STARTED
 (2/8) perf/perf_pmu.py:PerfBasic.test_config_PPC_RTAS:  PASS (0.54 s)
 (3/8) perf/perf_pmu.py:PerfBasic.test_cpu_event_count: STARTED
 (3/8) perf/perf_pmu.py:PerfBasic.test_cpu_event_count:  PASS (0.03 s)
 (4/8) perf/perf_pmu.py:PerfBasic.test_hv_24x7_event_count: STARTED
 (4/8) perf/perf_pmu.py:PerfBasic.test_hv_24x7_event_count:  PASS (0.03 s)
 (5/8) perf/perf_pmu.py:PerfBasic.test_imc_event_count: STARTED
 (5/8) perf/perf_pmu.py:PerfBasic.test_imc_event_count:  SKIP: This test is for PowerNV
 (6/8) perf/perf_pmu.py:PerfBasic.test_thread_event_count: STARTED
 (6/8) perf/perf_pmu.py:PerfBasic.test_thread_event_count:  SKIP: This test is for PowerNV
 (7/8) perf/perf_pmu.py:PerfBasic.test_nest_cpumask: STARTED
 (7/8) perf/perf_pmu.py:PerfBasic.test_nest_cpumask:  SKIP: This test is for PowerNV
 (8/8) perf/perf_pmu.py:PerfBasic.test_core_cpumask: STARTED
 (8/8) perf/perf_pmu.py:PerfBasic.test_core_cpumask:  SKIP: This test is for PowerNV

After the change:
 (1/8) perf/perf_pmu.py:PerfBasic.test_config_PMU_sysfs: STARTED
 (1/8) perf/perf_pmu.py:PerfBasic.test_config_PMU_sysfs:  PASS (0.37 s)
 (2/8) perf/perf_pmu.py:PerfBasic.test_config_PPC_RTAS: STARTED
 (2/8) perf/perf_pmu.py:PerfBasic.test_config_PPC_RTAS:  PASS (0.31 s)
 (3/8) perf/perf_pmu.py:PerfBasic.test_cpu_event_count: STARTED
 (3/8) perf/perf_pmu.py:PerfBasic.test_cpu_event_count:  PASS (0.03 s)
 (4/8) perf/perf_pmu.py:PerfBasic.test_hv_24x7_event_count: STARTED
 (4/8) perf/perf_pmu.py:PerfBasic.test_hv_24x7_event_count:  PASS (0.03 s)
 (5/8) perf/perf_pmu.py:PerfBasic.test_imc_event_count: STARTED
 (5/8) perf/perf_pmu.py:PerfBasic.test_imc_event_count:  SKIP: This test is for PowerNV
 (6/8) perf/perf_pmu.py:PerfBasic.test_thread_event_count: STARTED
 (6/8) perf/perf_pmu.py:PerfBasic.test_thread_event_count:  SKIP: This test is for PowerNV
 (7/8) perf/perf_pmu.py:PerfBasic.test_nest_cpumask: STARTED
 (7/8) perf/perf_pmu.py:PerfBasic.test_nest_cpumask:  SKIP: This test is for PowerNV
 (8/8) perf/perf_pmu.py:PerfBasic.test_core_cpumask: STARTED
 (8/8) perf/perf_pmu.py:PerfBasic.test_core_cpumask:  SKIP: This test is for PowerNV